# Conflicts:
#   .env

+++ app/qc_ingest/model/iqvsectionlock_db.py
@@ -1,6 +1,8 @@
 from sqlalchemy import Column, DateTime
 from datetime import datetime
 from .__base__ import SchemaBase, schema_to_dict, MissingParamException
+from ...crud.pd_user import CRUDUserSearch
+from ...models.pd_user import User
 from sqlalchemy.dialects.postgresql import TEXT
 
 
@@ -9,6 +11,7 @@ class IqvsectionlockDb(SchemaBase):
     link_id = Column(TEXT, primary_key=True, nullable=False)
     doc_id = Column(TEXT)
     userId = Column(TEXT)
+    user_name = Column(TEXT)
     last_updated = Column(DateTime(timezone=True),
                           default=datetime.utcnow, nullable=False)
 
@@ -25,11 +28,13 @@ class IqvsectionlockDb(SchemaBase):
         if not obj:
             data['section_lock'] = True
             data['userId'] = ''
+            data['user_name'] = ''
             data['last_updated'] = ''
         else:
             obj_dict = schema_to_dict(obj)
             data['section_lock'] = False
             data['userId'] = obj_dict['userId']
+            data['user_name'] = obj_dict['user_name']
             data['last_updated'] = obj_dict['last_updated']
         return data
 
@@ -43,13 +48,21 @@ class IqvsectionlockDb(SchemaBase):
         
         if data.get('section_lock') == False:
             section_info = IqvsectionlockDb()
-            section_info.link_id = data['link_id']
-            section_info.doc_id = data['doc_id']
-            section_info.userId = data['userId']
-            section_info.last_updated = datetime.utcnow()
+            crud_user_search = CRUDUserSearch(User)
+            section_info.link_id = data.get('link_id')
+            section_info.doc_id = data.get('doc_id','')
+            section_info.userId = user_id = data.get('userId','')
+            user_name_obj = crud_user_search.get_by_username(session, user_id)
+            if not user_name_obj:
+                raise MissingParamException("{0} user_id in User DB".format(user_id))
+            user_name = user_name_obj.first_name + ' ' + user_name_obj.last_name
+            section_info.user_name = user_name
+            section_info.last_updated = data['last_updated'] = datetime.utcnow()
             session.add(section_info)
         else:
             obj = session.query(IqvsectionlockDb).filter(
                 IqvsectionlockDb.link_id == data['link_id']).first()
+            if not obj:
+                raise MissingParamException("{0} in iqv section lock DB".format(data['link_id']))
             session.delete(obj)
         return data
